fix bug all CRUD

+++ 00-java-spring-soundcloud/src/main/java/vn/bachdao/soundcloud/service/TrackService.java
@@ -2,7 +2,6 @@ package vn.bachdao.soundcloud.service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -10,10 +9,8 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import vn.bachdao.soundcloud.domain.Comment;
-import vn.bachdao.soundcloud.domain.Genre;
 import vn.bachdao.soundcloud.domain.Track;
 import vn.bachdao.soundcloud.repository.CommentRepository;
-import vn.bachdao.soundcloud.repository.GenreRepository;
 import vn.bachdao.soundcloud.repository.TrackRepository;
 import vn.bachdao.soundcloud.service.dto.repsonse.ResultPaginationDTO;
 import vn.bachdao.soundcloud.service.dto.repsonse.track.ResGetTrackDTO;
@@ -22,28 +19,20 @@ import vn.bachdao.soundcloud.service.mapper.TrackMapper;
 @Service
 public class TrackService {
     private final TrackRepository trackRepository;
-    private final GenreRepository genreRepository;
+
     private final CommentRepository commentRepository;
     private final TrackMapper trackMapper;
 
     public TrackService(
             TrackRepository trackRepository,
             TrackMapper trackMapper,
-            GenreRepository genreRepository,
             CommentRepository commentRepository) {
         this.trackRepository = trackRepository;
         this.trackMapper = trackMapper;
-        this.genreRepository = genreRepository;
         this.commentRepository = commentRepository;
     }
 
     public Track handleCreateTrack(Track track) {
-        if (track.getGenres() != null) {
-            List<Long> ids = track.getGenres().stream().map(item -> item.getId()).collect(Collectors.toList());
-            List<Genre> dbGenres = this.genreRepository.findByIdIn(ids);
-            track.setGenres(dbGenres);
-        }
-
         return this.trackRepository.save(track);
     }
 
@@ -68,8 +57,10 @@ public class TrackService {
         currentTrack.getPlaylists().forEach(playlist -> playlist.getTracks().remove(currentTrack));
 
         // delete track from comment table
-        List<Comment> comments = currentTrack.getComments();
-        this.commentRepository.deleteAll(comments);
+        if (currentTrack.getComments() != null) {
+            List<Comment> comments = currentTrack.getComments();
+            this.commentRepository.deleteAll(comments);
+        }
 
         // delete track from track table
         this.trackRepository.deleteById(id);
